main_windows_test.go: Make tests for Windows
